fix atualiza int

+++ Trabalho-01/pessoal/DoisCaraNumaMoto.h
@@ -10,7 +10,7 @@ public:
 
     DoisCaraNumaMoto();
     void desenha();
-    void atualiza(float a){}
+    void atualiza(int a){}
     void camera() {}
 };
