+++ Trabalho-01/pessoal/Personagem.h
@@ -11,7 +11,7 @@ public:
 
     Personagem();
     void desenha();
-    void atualiza(float a){}
+    void atualiza(int a){}
     void camera();
 };
